chuyển đơn vị, thay empty chart, thay đổi trường export, thêm pop up edit

+++ lib/models/sugar_info/sugar_info.dart
@@ -34,8 +34,8 @@ class Conditions {
 
 class SugarAmount {
   String? status;
-  int? minValue;
-  int? maxValue;
+  double? minValue;
+  double? maxValue;
 
   SugarAmount({this.status, this.minValue, this.maxValue});
